Adjust test to be like other C tests
