+++ test/types/cptr/chplClassInC/.gitignore
@@ -1,0 +1,2 @@
+# Chapel library header and archive
+lib-dir/
