+++ test/types/cptr/takeClassToCAndBack/CLEANFILES
@@ -1,0 +1,3 @@
+lib-dir/liblib.a
+lib-dir/lib.h
+lib-dir/liblib.h
