--- test/types/cptr/takeClassToCAndBack/Makefile
@@ -1,4 +1,0 @@
-include lib/Makefile.takeClassToCAndBack
-
-takeClassToCAndBack: takeClassToCAndBack.c lib/takeClassToCAndBack.a
-	$(CHPL_COMPILER) $(CHPL_CFLAGS) -o takeClassToCAndBack takeClassToCAndBack.c $(CHPL_LDFLAGS)
