+++ test/types/cptr/takeClassToCAndBack/lib.compopts
@@ -1,0 +1,1 @@
+-s cPtrToLogicalValue=true
