--- test/types/cptr/takeClassToCAndBack/lib/Makefile.takeClassToCAndBack
@@ -1,17 +1,0 @@
-CHPL_RUNTIME_LIB = /Users/rift/chapel/lib
-
-CHPL_RUNTIME_INCL = /Users/rift/chapel/runtime/include
-
-CHPL_THIRD_PARTY = /Users/rift/chapel/third-party
-
-CHPL_HOME = /Users/rift/chapel
-
-CHPL_CFLAGS = -Ilib -Wno-unused -Wno-uninitialized -Wno-pointer-sign -Wno-incompatible-pointer-types -Wno-tautological-compare -I/Users/rift/chapel/modules/internal -I$(CHPL_RUNTIME_INCL)/localeModels/flat -I$(CHPL_RUNTIME_INCL)/localeModels -I$(CHPL_RUNTIME_INCL)/comm/none -I$(CHPL_RUNTIME_INCL)/comm -I$(CHPL_RUNTIME_INCL)/tasks/fifo -I$(CHPL_RUNTIME_INCL)/. -I$(CHPL_RUNTIME_INCL)/./qio -I$(CHPL_RUNTIME_INCL)/./atomics/cstdlib -I$(CHPL_RUNTIME_INCL)/./mem/cstdlib -I$(CHPL_THIRD_PARTY)/utf8-decoder -I. -I/opt/homebrew/include
-
-CHPL_LDFLAGS = -Llib -ltakeClassToCAndBack -L$(CHPL_RUNTIME_LIB)/darwin/llvm/arm64/cpu-native/loc-flat/comm-none/tasks-fifo/tmr-generic/unwind-none/mem-cstdlib/atomics-cstdlib/hwloc-none/re2-none/fs-none/lib_pic-none/san-none -lchpl -lm -lpthread -L/opt/homebrew/lib -ltakeClassToCAndBack
-
-CHPL_COMPILER = /opt/homebrew/Cellar/llvm@15/15.0.7/bin/clang
-
-CHPL_LINKER = /opt/homebrew/Cellar/llvm@15/15.0.7/bin/clang++
-
-CHPL_LINKERSHARED = /opt/homebrew/Cellar/llvm@15/15.0.7/bin/clang++ -shared
