--- test/types/cptr/takeClassToCAndBack/lib/takeClassToCAndBack.h
@@ -1,9 +1,0 @@
-#include "stdchpl.h"
-#include "ctype.h"
-#include "wctype.h"
-void chpl__init_takeClassToCAndBack(int64_t _ln,
-                                    int32_t _fn);
-void getInstance(c_void_ptr * addrOut,
-                 int64_t arg);
-int64_t useInstance(c_void_ptr instancePtr);
-void deleteInstance(c_void_ptr instancePtr);
