+++ test/types/cptr/takeClassToCAndBack/takeClassToCAndBack.good
@@ -1,0 +1,1 @@
+4
