+++ test/types/cptr/takeClassToCAndBack/takeClassToCAndBack.precomp
@@ -1,0 +1,2 @@
+#!/usr/bin/env bash
+$3 -s cPtrToLogicalValue=true --library --library-dir lib-dir -o liblib lib.chpl
